refactor create verbwire lib for requests
